requirements.txt
Python 3.9+

OpenCV

Streamlit

NumPy

pandas

tqdm

Pillow